back-master/requirements.txt
# Core dependencies
Flask==3.1.0
flask-cors==4.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==21.2.0
pymongo==4.6.2
blinker==1.9.0

# Document processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pillow>=9.0.0
xlrd==2.0.1
openpyxl==3.1.2
pandas>=1.5.0
tabula-py==2.9.0
PyMuPDF>=1.23.0 # For robust PDF text/image extraction


# AI and ML
huggingface-hub>=0.23.0,<1.0
# transformers==4.41.1 # Temporarily commented out due to build issues (requires Rust for tokenizers)
openai>=1.68.2 # Updated to resolve conflict with langchain-openai
mistralai==1.0.1
langchain-deepseek==0.1.3
langchain>=0.2.0 # Updated to resolve langchain-core conflict with langchain-deepseek
langchain-community>=0.2.0 # Updated to resolve langchain-core conflict
tiktoken>=0.7 # Updated to resolve conflict with langchain-openai
# spacy==3.7.4 # Temporarily commented out due to build issues (requires C compiler for blis)
# sentence-transformers==2.3.1 # Temporarily commented out (depends on transformers)

# Data handling
numpy==1.26.4
# scipy==1.13.0 # Temporarily commented out due to build issues (requires Fortran compiler)
nltk==3.8.1
beautifulsoup4==4.12.3
# lxml==5.2.0 # Temporarily commented out due to build issues (requires libxml2/libxslt system libraries)

# Testing
pytest==8.3.5
pytest-flask==1.3.0
pytest-cov==6.0.0

# Development tools
black==24.3.0
isort==5.12.0
flake8==7.0.0

# Document image processing
poppler-utils==0.1.0
Wand==0.6.13
opencv-python-headless
paddleocr

# Security
cryptography==42.0.8
PyJWT==2.8.0

# AWS integration (optional)
boto3==1.34.71
awscli==1.32.71

# Task Queue
celery[redis]==5.4.0

# Payment processing
# paddle-python==0.0.5 # Temporarily commented out due to version not found
